requirements.txt
openai
python-dotenv
requests
pydantic
PyPDF2
